requirements.txt
pandas==2.2.0
numpy==1.26.4
scikit-learn==1.4.0
lightgbm==4.3.0
sentence-transformers==2.3.1
pinecone-client==3.0.0
langchain==0.1.0
langchain-community==0.0.10
langchain-openai==0.0.2.post1
fastapi==0.109.0
uvicorn==0.27.0.post1
jupyter==1.0.0
matplotlib==3.8.2
seaborn==0.13.2
tqdm==4.66.1
kaggle==1.6.5
ipywidgets
ipykernel
pyarrow
requests
